Move columns controller writes to Carto
